WIP implementing controller for transactions
