+++ src/routes/transactions.js
@@ -1,59 +1,15 @@
 import { Router } from "express";
 import { convertBytes16ToUUID, convertUUIDToBytes16 } from "../utils/convertUUID.js";
 import { toScaledBigNumber } from "../utils/convertToFixedPointDecimals.js";
+import { convertAndCreateIssuanceStockOnchain } from "../db/controllers/transactions/issuanceController.js";
 
 const transactions = Router();
 
 transactions.post("/issuance/stock", async (req, res) => {
     const { contract } = req;
-    const {
-        stakeholderId,
-        stockClassId,
-        quantity,
-        sharePrice,
-        stockPlanId,
-        shareNumbersIssued,
-        vestingTermsId,
-        costBasis,
-        stockLegendIds,
-        issuanceType,
-        comments,
-        customId,
-        boardApprovalDate,
-        stockholderApprovalDate,
-        considerationText,
-        securityLawExemptions,
-    } = req.body;
-
-    const stakeholderIdBytes16 = convertUUIDToBytes16(stakeholderId);
-    const stockClassIdBytes16 = convertUUIDToBytes16(stockClassId);
-    const vestingTermsBytes16 = convertUUIDToBytes16(vestingTermsId);
-    const stockPlanIdBytes16 = convertUUIDToBytes16(stockPlanId);
-
-    const quantityScaled = toScaledBigNumber(quantity);
-    const sharePriceScaled = toScaledBigNumber(sharePrice);
 
     try {
-        const tx = await contract.issueStockByTA(
-            stockClassIdBytes16,
-            stockPlanIdBytes16,
-            shareNumbersIssued, // not converted
-            sharePriceScaled,
-            quantityScaled,
-            vestingTermsBytes16,
-            costBasis, // not converted
-            stockLegendIds, // not converted
-            issuanceType,
-            comments,
-            customId,
-            stakeholderIdBytes16,
-            boardApprovalDate,
-            stockholderApprovalDate,
-            considerationText,
-            securityLawExemptions
-        );
-        await tx.wait();
-        console.log("Issued stock successfully");
+        await convertAndCreateIssuanceStockOnchain(contract, req.body);
 
         res.status(200).send("success");
     } catch (error) {
